#2 - Refactor: move class to its package
