+++ src/main/kotlin/com/gmaur/investment/robotadvisor/infrastructure/RebalanceRequest.kt
@@ -1,4 +1,4 @@
-package com.gmaur.investment.robotadvisor
+package com.gmaur.investment.robotadvisor.infrastructure
 
 import com.gmaur.investment.robotadvisor.domain.AssetAllocation
 import com.gmaur.investment.robotadvisor.domain.Portfolio
